requirements.txt
streamlit
openai
paddleocr
pdfplumber
Pillow
python-docx
pandas
numpy
